draw floor and roof

+++ mango3d/Source/mango3d/Entity/RoomData.cpp
@@ -47,7 +47,7 @@ void RoomData::UpdateInitData(std::map<FString, WallData*>& wallDataMap) {
 
 std::vector<CornerData*> RoomData::GetCorners()const {
   std::vector<CornerData*> result;
-  for (int i = 0; i < walls_.size(); i++) {
+  for (size_t i = 0; i < walls_.size(); i++) {
     int j = (i + 1) % walls_.size();
     WallData* wall1 = walls_[i];
     WallData* wall2 = walls_[j];
